Memory block number and EEPROM updates

+++ src/commands.h
@@ -81,7 +81,7 @@ extern cmd_t reset_cmd;
 void finish_current_cmd(bool succeeded);
 
 void auto_data_col_timer_cb(void);
-void populate_header(mem_header_t* header, uint8_t block_num, uint8_t error);
+void populate_header(mem_header_t* header, uint32_t block_num, uint8_t error);
 
 void append_header_to_tx_msg(mem_header_t* header);
 void append_fields_to_tx_msg(uint32_t* fields, uint8_t num_fields);
